requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cssselect2==0.7.0
distro==1.9.0
elevenlabs==1.7.0
exceptiongroup==1.2.2
fonttools==4.53.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
Markdown==3.7
numpy==2.1.0
openai==1.43.0
pillow==10.4.0
pvcobra==2.0.2
pvporcupine==3.0.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pydyf==0.11.0
pygame==2.6.0
pyphen==0.16.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
tinycss2==1.3.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
websockets==13.0.1
zopfli==0.2.3
